backend/pyproject.toml
[tool.poetry]
name = "todo_list_app_backend"
version = "0.1.0"
description = ""
authors = ["Sami Sarsa <sami.j.sarsa@jyu.fi>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.11"}
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
databases = { extras = ["asyncpg"], version = "^0.9.0" }
passlib = "^1.7.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
asgi-lifespan = "^2.1.0"
async-asgi-testclient = "^1.4.11"
websockets = "^15.0"


[tool.poetry.group.dev.dependencies]
asgi-lifespan = "^2.1.0"
black = "^24.2.0"
httpx = "^0.27.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
testcontainers = "^4.3.3"
psycopg = "^3.1.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.pytest.ini_options]
pythonpath = ["."]

[mutmut]
paths_to_mutate = "."
